# optimize generic_bfs_edges function