@ImHereForTheCookies 2023-02-09
Makes sense to me. Related to [this pull](https://github.com/networkx/networkx/pull/6366) for someone who has the power to merge.